+++ src/main/java/com/uisek/servicio/admin/model/service/UsuarioService.java
@@ -8,9 +8,9 @@ import com.uisek.servicio.admin.model.entity.Usuario;
 public interface UsuarioService {
 
 	public List<Usuario> findAll();
-	public Usuario findById(Long id);
+	public Usuario findById(Integer id);
 	public Usuario save(Usuario entidad);
-	public void deleteById(Long id);
+	public void deleteById(Integer id);
 	
 	public Usuario findByNombreUsuario(String nombreUsuario);
 	public Usuario findByNombreUsuarioAndClave(String nombreUsuario, String clave);
